[refactor] Move `GridViewModel` to single file
